Update vignette to work with new tasic2016data
A few columns in tasic2016data has been renamed since the vignette was
made. This commit puts the vignette back into a runnable state and
includes a new knitted HTML file. Note that it does not replicate the
originally results exactly, particularly past the t-SNE plotting
section. This may indicate that further changes are needed to make a
current vignette.